+++ packages/cortex-core/src/components/tag/tag.ts
@@ -1,5 +1,5 @@
-import { createTV } from 'tailwind-variants';
-import { fontSize } from '@tecsinapse/cortex-core';
+import { ClassProp, createTV, VariantProps } from 'tailwind-variants';
+import { fontSize } from '../../tokens/definitions';
 
 const myTV = createTV({
   twMergeConfig: {
@@ -10,7 +10,7 @@ const myTV = createTV({
   },
 });
 
-export const tag = myTV({
+const tagStyles = myTV({
   base: 'rounded-micro px-micro py-nano w-fit font-bold text-label items-center flex',
   variants: {
     intent: {
@@ -25,3 +25,14 @@ export const tag = myTV({
     intent: 'secondary',
   },
 });
+
+export type TagVariants = VariantProps<typeof tagStyles> & ClassProp;
+
+/**
+ * Represents the tag component with specified variants.
+ * @param {TagVariants} props - The properties for the tag component.
+ * @param {string=} [props.intent=secondary] -  The intent variant for the tag (e.g., success, primary, secondary, info, white).
+ * @param {string=} [props.className] - The additional CSS classes for the tag.
+ * @param {string=} [props.class] - The additional CSS classes for the tag.
+ */
+export const tag = (props: TagVariants) => tagStyles(props);