+++ packages/react-lab/src/select.stories.tsx
@@ -2,9 +2,9 @@ import React, { useState } from 'react';
 import { StoryFn } from '@storybook/react';
 import {
   buttonSelect,
-  dropDown,
+  dropdown,
   option as styleOption,
-} from '@tecsinapse/cortex-core';
+} from '@tecsinapse/cortex-core/src/components';
 export default {
   title: 'Lab/Select',
   component: <div />,
@@ -33,7 +33,7 @@ const Template: StoryFn = () => {
         <p>{value ? value : 'Select one'}</p>
         <div>\/</div>
       </button>
-      <ul role={'select'} className={dropDown({ open, className: 'absolute' })}>
+      <ul role={'select'} className={dropdown({ open, className: 'absolute' })}>
         {options.map(option => (
           <li
             role={'option'}