+++ packages/react-core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.3.5](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-core@3.3.3...@tecsinapse/react-core@3.3.5) (2024-02-07)
+
+**Note:** Version bump only for package @tecsinapse/react-core
+
+## [3.3.4](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-core@3.3.2...@tecsinapse/react-core@3.3.4) (2024-01-15)
+
+**Note:** Version bump only for package @tecsinapse/react-core
+
 ## [3.3.3](https://github.com/tecsinapse/design-system/compare/@tecsinapse/react-core@3.3.1...@tecsinapse/react-core@3.3.3) (2023-12-19)
 
 **Note:** Version bump only for package @tecsinapse/react-core